Refactoring on the way
