+++ Cargo.toml
@@ -34,6 +34,7 @@ bitflags = {version = "1.3.2", default-features = false}
 clap = "4.1.6"
 env_logger = "0.10.0"
 futures = {version = "0.3.21", default-features = false}
+hash-db = { version = "0.16.0", default-features = false }
 hex = {version = "0.4.0", default-features = false}
 hex-literal = "0.3.4"
 jsonrpsee = "0.16.2"
