+++ crates/kilt-dip-support/Cargo.toml
@@ -11,6 +11,9 @@ repository.workspace = true
 version.workspace = true
 
 [dependencies]
+# External dependencies
+hash-db.workspace = true
+
 # Internal dependencies
 did.workspace = true
 dip-support.workspace = true
@@ -41,6 +44,7 @@ sp-io = { workspace = true, features = ["std"] }
 [features]
 default = ["std"]
 std = [
+  "hash-db/std",
   "did/std",
   "dip-support/std",
   "pallet-dip-consumer/std",
